requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2

# Bybit API
pybit==5.7.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Technical indicators
pandas-ta==0.3.14b0
ta==0.11.0

# Backtesting
backtrader==1.9.78.123
vectorbt==0.26.2

# GUI
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2

# Async support
asyncio==3.4.3
websockets==12.0
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Type hints
typing-extensions==4.9.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
